time fixed

+++ philo.h
@@ -6,7 +6,7 @@
 /*   By: msoklova <msoklova@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/02 14:29:50 by msoklova          #+#    #+#             */
-/*   Updated: 2024/11/05 13:06:22 by msoklova         ###   ########.fr       */
+/*   Updated: 2024/11/06 16:41:51 by msoklova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ void		*death_monitor(void *arg);
 
 /* UTILS */
 int			ft_atoi(const char *str);
-long		curr_time(void);
+long long		curr_time(void);
 void		print_action(t_events *events, int id, const char *action);
 
 #endif
